Update instance resize method to support waiting and package name and shortId in addition to id.
I also took the liberty to change the internal _waitForInstanceName method to _waitForInstanceUpdate. I've basically abstracted the check, so now instead of passing in a name, you pass in a function which can do the check and return whether or not the update has been applied.

This was so I could reuse this function to check for the resize update as well.